Move data fetching to route
